New version of the GCodeProccessor complete
- updated version of the GCodeProcessor
- create as iterable
- the iterator return elementary operations to perform
- simulation is run outsite the class
