+++ tests/test_g_code_processor.py
@@ -1,6 +1,5 @@
 import os
 import unittest
-import time
 import numpy as np
 import pybullet as p
 import pybullet_data
@@ -8,19 +7,18 @@ import pybullet_industrial as pi
 
 
 def couple_endeffector(gripper: pi.Gripper, robot: pi.RobotBase, link: chr):
-
     gripper.couple(robot, link)
-    for _ in range(25):
-        p.stepSimulation()
-        time.sleep(0.01)
 
 
 def decouple_endeffector(gripper: pi.Gripper):
-
     gripper.decouple()
-    for _ in range(25):
-        p.stepSimulation()
-        time.sleep(0.01)
+
+
+def run_simulation(processor: pi.GCodeProcessor):
+    for _ in processor:
+        for _ in range(20):
+            for _ in range(10):
+                p.stepSimulation()
 
 
 def check_robot_position(robot: pi.RobotBase, target_position,
@@ -60,19 +58,6 @@ def check_tool_position(endeffetor: pi.EndeffectorTool, target_position,
     return within_precision
 
 
-def create_command(gcode_object: pi.GCodeProcessor, commands):
-
-    text_gcode = "text_gcode.txt"
-
-    with open(text_gcode, "w", encoding="utf-8") as f:
-        f.writelines(commands)
-
-    gcode = gcode_object.read_gcode(text_gcode)
-    os.remove(text_gcode)
-
-    return gcode
-
-
 class Test_GCodeProcessor(unittest.TestCase):
 
     def test_read_gcode(self):
@@ -81,28 +66,28 @@ class Test_GCodeProcessor(unittest.TestCase):
         cmd1 = "% This is a comment\n"
         cmd2 = "G0 X1.0 Y2.0\n"
         cmd3 = "G1 X2.0 Y3.0\n"
-        cmd4 = "G2 X3.0 Y4.0 Z5.0\n"
-        commands = [cmd1, cmd2, cmd3, cmd4]
+        cmd4 = "G2 X3.0 Y4.0 Z5.0"
+        g_code_input = cmd1 + cmd2 + cmd3 + cmd4
 
         # Creating the G-Code Object
-        gcode_object = pi.GCodeProcessor()
-
-        gcode = create_command(gcode_object, commands)
+        test_object = pi.GCodeProcessor(g_code_input)
 
         # Check that the length of the gcode list is correct
-        self.assertEqual(len(gcode), 3)
+        self.assertEqual(len(test_object.gcode), 3)
 
         # Check that the first line is correct
-        self.assertEqual(gcode[0], [['G', 0], ['X', 1.0], ['Y', 2.0]])
+        self.assertEqual(test_object.gcode[0], [
+                         ['G', 0], ['X', 1.0], ['Y', 2.0]])
 
         # Check that the second line is correct
-        self.assertEqual(gcode[1], [['G', 1], ['X', 2.0], ['Y', 3.0]])
+        self.assertEqual(test_object.gcode[1], [
+                         ['G', 1], ['X', 2.0], ['Y', 3.0]])
 
         # Check that the third line is correct
-        self.assertEqual(
-            gcode[2], [['G', 2], ['X', 3.0], ['Y', 4.0], ['Z', 5.0]])
+        self.assertEqual(test_object.gcode[2], [['G', 2], [
+                         'X', 3.0], ['Y', 4.0], ['Z', 5.0]])
 
-    def test_run_gcode(self):
+    def test_simulation(self):
 
         dirname = os.path.dirname(__file__)
         parentDir = os.path.dirname(dirname)
@@ -138,7 +123,7 @@ class Test_GCodeProcessor(unittest.TestCase):
             p.stepSimulation()
 
         dirname = os.path.dirname(__file__)
-        test_object = pi.GCodeProcessor(robot, interpolation_steps=100)
+        test_object = pi.GCodeProcessor(robot=robot, interpolation_steps=100)
 
         pos1 = np.array([1.9, -0.5, 1.5])
         ori1 = np.array([-1.5079, 0.0, 0.0])
@@ -150,28 +135,30 @@ class Test_GCodeProcessor(unittest.TestCase):
         ori_precision = 0.04
 
         # Test 1: G0 without coupled tool
-        cmd1 = "G0 X1.9 Y-0.5 Z1.5 A-1.5079 B0 C0"
-        commands = [cmd1]
-        g_code = create_command(test_object, commands)
-        test_object.run_gcode(g_code)
+        command = "G0 X1.9 Y-0.5 Z1.5 A-1.5079 B0 C0"
+        test_object.gcode = test_object.read_gcode(command)
+        run_simulation(test_object)
+
         self.assertTrue(check_robot_position(
             robot, pos1, ori1, pos_precision, ori_precision))
 
         # Test 2: G1 without coupled tool
         cmd1 = "G1 X2.2\n"
         cmd2 = "G1 Y-0.3 Z1.6 A-1.403"
-        commands = [cmd1, cmd2]
-        g_code = create_command(test_object, commands)
-        test_object.run_gcode(g_code)
+        commands = cmd1+cmd2
+        test_object.gcode = test_object.read_gcode(commands)
+        run_simulation(test_object)
+
         self.assertTrue(check_robot_position(
             robot, pos2, ori2, pos_precision, ori_precision))
 
         # Test 3: G54
         cmd1 = "G54\n"
         cmd2 = "G1 X-0.3 Y-0.2 Z-0.1 A-0.1049"
-        commands = [cmd1, cmd2]
-        g_code = create_command(test_object, commands)
-        test_object.run_gcode(g_code)
+        commands = cmd1 + cmd2
+        test_object.gcode = test_object.read_gcode(commands)
+        run_simulation(test_object)
+
         self.assertTrue(check_robot_position(
             robot, pos1, ori1, pos_precision, ori_precision))
 
@@ -180,9 +167,10 @@ class Test_GCodeProcessor(unittest.TestCase):
         # Test 4: G500
         cmd1 = "G500\n"
         cmd2 = "G1 X2.2 Y-0.3 Z1.6"
-        commands = [cmd1, cmd2]
-        g_code = create_command(test_object, commands)
-        test_object.run_gcode(g_code)
+        commands = cmd1 + cmd2
+        test_object.gcode = test_object.read_gcode(commands)
+        run_simulation(test_object)
+
         self.assertTrue(check_robot_position(
             robot, pos2, ori2, pos_precision, ori_precision))
 
@@ -193,9 +181,10 @@ class Test_GCodeProcessor(unittest.TestCase):
         # Test 5: G2 with G19
         cmd1 = "G19\n"
         cmd2 = "G2 X1.9 Y-0.5 Z1.5 R1"
-        commands = [cmd1, cmd2]
-        g_code = create_command(test_object, commands)
-        test_object.run_gcode(g_code)
+        commands = cmd1 + cmd2
+        test_object.gcode = test_object.read_gcode(commands)
+        run_simulation(test_object)
+
         self.assertTrue(check_robot_position(
             robot, pos1, ori1, pos_precision, ori_precision))
 
@@ -203,9 +192,10 @@ class Test_GCodeProcessor(unittest.TestCase):
         cmd1 = "G18\n"
         cmd2 = "G3 X2.2 Y-0.3 Z1.6 R1\n"
         cmd3 = "G17"
-        commands = [cmd1, cmd2, cmd3]
-        g_code = create_command(test_object, commands)
-        test_object.run_gcode(g_code)
+        commands = cmd1 + cmd2 + cmd3
+        test_object.gcode = test_object.read_gcode(commands)
+        run_simulation(test_object)
+
         self.assertTrue(check_robot_position(
             robot, pos2, ori2, pos_precision, ori_precision))
 
@@ -232,16 +222,17 @@ class Test_GCodeProcessor(unittest.TestCase):
             test_gripper, robot, 'link6'))
 
         # Adapt Test Object
-        test_object = pi.GCodeProcessor(robot, endeffector_list, m_commands,
-                                        t_commands)
+        test_object.endeffector_list = endeffector_list
+        test_object.m_commands = m_commands
+        test_object.t_commands = t_commands
 
         # Test 7: T1/M command and G0 command with tool coupled
         cmd1 = "T1\n"
         cmd2 = "G0 X1.9 Y-0.5 Z1.5 A-3.1 B0 C0\n"
         cmd3 = "M1"
-        commands = [cmd1, cmd2, cmd3]
-        g_code = create_command(test_object, commands)
-        test_object.run_gcode(g_code)
+        commands = cmd1 + cmd2 + cmd3
+        test_object.gcode = test_object.read_gcode(commands)
+        run_simulation(test_object)
         self.assertTrue(test_gripper.is_coupled())
         self.assertTrue(check_tool_position(test_gripper, pos1,
                         ori1, pos_precision, ori_precision))
@@ -250,9 +241,9 @@ class Test_GCodeProcessor(unittest.TestCase):
         # Test 8: G1 command with tool coupled and T0 command for decoupling
         cmd1 = "G1 X2.2 Y-0.3 Z1.6\n"
         cmd2 = "T0 "
-        commands = [cmd1, cmd2]
-        g_code = create_command(test_object, commands)
-        test_object.run_gcode(g_code)
+        commands = cmd1 + cmd2
+        test_object.gcode = test_object.read_gcode(commands)
+        run_simulation(test_object)
         self.assertTrue(not test_gripper.is_coupled())
         self.assertTrue(check_tool_position(test_gripper, pos2,
                         ori2, pos_precision, ori_precision))
